apps/spotify_dashboard/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
databricks-sdk>=0.20.0

